==> 01-data.Rmd <==
# Obtaining the data

```{r, include=FALSE}
knitr::opts_chunk$set(
  echo=TRUE, eval=FALSE, warning=FALSE, message=FALSE,
  class.source="language-r", class.output="language-r"
)
```

## Packages

```{r}
library(rvest)
library(dplyr)
library(purrr)
```

## Create function to get data for each year

### Try 2021

```{r}
url2021 <- "https://www.lottomaxnumbers.com/numbers/2021"

nums2021 <- read_html(url2021) %>%
  html_element(css="table") %>%
  html_table(header=TRUE)
```

### General case

```{r}
getnums <- function(year){
  yearurl <- paste0("https://www.lottomaxnumbers.com/numbers/", year)
  
  read_html(yearurl) %>%
    html_element(css="table") %>%
    html_table(header=TRUE)
}
```

## Get the data

Just going to row-bind it all.

```{r}
input_years <- 2019:2021 %>%
  set_names(., nm=.)

lottomax <- map_dfr(input_years, getnums, .id="year")
```

```{r}
saveRDS(lottomax, "./data/lottomax.RDS", version=2)
```
