==> 02-clean.Rmd <==
# Cleaning the data

```{r, include=FALSE}
knitr::opts_chunk$set(
  echo=TRUE, message=FALSE, warning=FALSE,
  class.source="language-r", class.output="language-r"
)
```

## Packages

```{r}
library(tidyverse)
library(lubridate)
```

## Load the data

```{r}
lottomax <- readRDS("./data/lottomax.RDS")

lottomax

as.character(lottomax[1,2])

as.character(lottomax[1,3])
```

## Clean the data

To do:

- Rename (shorten) `Date` and `Draw Results` columns using `tidyselect::contains` since column
names were recently updated to something similar but not the same. I.e. don't use a fixed old_name 
in `new_name = old_name` convention
- Remove `Jackpot` column
- Remove rows mentioning "Next Lotto Max Jackpot ...". This needs to be done before parsing dates

- `date`: remove the non-space characters and the "With Max Millions!" phrase
- Need to filter to only include dates after May 14, 2019 (inclusive) since LottoMax
was changed to now have 50 possible numbers instead of 49
- `result`: remove the non-space characters and reformat this column so that the numbers
are easy to access and manipulate
- Remember that the last number is a "bonus" number
- Preferable to sort data too

```{r}
lottomax_clean <- lottomax %>%
  rename("date" = contains("date")) %>%
  rename("result" = contains("results")) %>%
  # Convert `year` to factor
  mutate(across(year, as.factor)) %>%
  # Keep rows that *don't* contain "Next Lotto Max Jackpot"
  filter(str_detect(date, pattern="Next", negate=TRUE)) %>%
  # Repair `date` to be an actual date
  mutate(
    # Remove excessive whitespace from `date` column
    date = str_squish(date),
    # Remove extra text
    date = str_remove(date, pattern="With Max Millions!"),
    # Parse date text in mm/dd/yyyy format
    date = mdy(date)
  ) %>%
  # Keep only rows that were drawn after May 14, 2019 (inclusive)
  filter(date >= mdy("May 14 2019")) %>%
  # Remove excessive whitespace from `result` column
  mutate(result = str_squish(result)) %>%
  select(year, date, result) %>%
  arrange(date)

lottomax_clean
```

A wide and long version of the data set will likely be useful.

- Wide version for analysis of "number positions"
- Long version for regular numerical summaries

### Wide version

```{r}
lottomax_wide <- lottomax_clean %>%
  # Separate all numbers into their own columns
  # Set `remove=FALSE` to check if we did it correctly
  separate(result, into=c(glue::glue("num{1:7}"), "bonus"), sep=" ", remove=FALSE) %>%
  # Conversion of strings to numbers that we can work with
  mutate(across(c(num1:num7, bonus), parse_integer)) 

lottomax_wide
```

```{r, eval=FALSE}
saveRDS(lottomax_wide, "./data/lottomax_wide.RDS", version=2)
```

### Long version

```{r}
lottomax_long <- lottomax_wide %>%
  pivot_longer(
    cols = c(num1:num7, bonus),
    names_to = "role",
    values_to = "number"
  )

lottomax_long
```

```{r, eval=FALSE}
saveRDS(lottomax_long, "./data/lottomax_long.RDS", version=2)
```
