==> 03-summaries.Rmd <==
# Numerical and graphical summaries

```{r, include=FALSE}
knitr::opts_chunk$set(
  echo=TRUE, message=FALSE, warning=FALSE,
  class.source="language-r", class.output="language-r",
  fig.align="center", fig.width=9
)
```


## Packages

```{r}
library(tidyverse)
library(lubridate)
theme_set(theme_bw())
```


## Plan

From the wide data:

- Most frequent minimum (`num1`), most frequent maximum (`num7`)
- Summary statistics for the minimum and maximum
- Distribution of the sum of the seven main numbers
- Sum of the seven main numbers with respect to time

From the long data:

- For the sake of illustration, repeat everything that was done for the wide data
- Distribution of numbers for each slot, i.e. `num1`, `num2`, ..., `bonus`
- Most frequent numbers in a given month
- Co-occurrences


## Wide data

```{r}
lottomax_wide <- readRDS("./data/lottomax_wide.RDS")

lottomax_wide
```


### Most frequent minimums and maximums - overall

```{r}
min_counts <- lottomax_wide %>%
  count(num1) %>%
  arrange(desc(n))

head(min_counts)
```

The drawn numbers should be treated as factors when plotting since these are discrete values. Note
that they are converted to factor when plotting, but we are not modifying the supplied data. If 
the drawn numbers were instead treated as continuous values, the axis would start at zero which just
look "off".

The bar graph will make use of `geom_col` rather than `geom_bar` since we have the counts already.
`geom_bar` would be used when we need to obtain counts from the data before making the bars.

```{r}
ggplot(min_counts, aes(x=factor(num1), y=n))+
  geom_col()+
  labs(x="Minimum number", y="Count")
```

Although it would be preferable to reduce some of the vertical gridlines, we should note that some 
numbers have actually never occurred as the minimum, i.e. the numbers appearing on the x-axis are not
consecutive. If we wish to reduce some of the gridlines, we need to make sure that all numbers 
from 1 to 50 are appearing on the x-axis.

We can fill in the unobserved numbers using the `complete` function found within the `tidyr` package.

```{r}
min_counts_complete <- min_counts %>%
  complete(
    num1 = full_seq(1:50, period=1),
    fill = list(n = 0)
  )
```

```{r}
ggplot(min_counts_complete, aes(x=factor(num1), y=n))+
  geom_col()+
  labs(x="Minimum number", y="Count")
```

Now reusing the same graph as above, let's reduce the number of vertical gridlines.

```{r}
ggplot(min_counts_complete, aes(x=factor(num1), y=n))+
  geom_col()+
  scale_x_discrete(breaks=factor(seq(1, 50, 7)))+
  labs(x="Minimum number", y="Count")
```

Realistically, the largest possible minimum number that could occur would be 44, resulting
in a draw of seven consecutive numbers (44 to 50). This means that we could actually shorten the
x-axis to only go up to 44.

Now let's repeat the same visulisation process for the maximum number of each draw.

```{r}
max_counts <- lottomax_wide %>%
  count(num7) %>%
  arrange(desc(n))

head(max_counts)
```

```{r}
ggplot(max_counts, aes(x=factor(num7), y=n))+
  geom_col()+
  labs(x="Maximum number", y="Count")
```

```{r}
max_counts_complete <- max_counts %>%
  complete(
    num7 = full_seq(1:50, period=1),
    fill = list(n = 0)
  )
```

```{r}
ggplot(max_counts_complete, aes(x=factor(num7), y=n))+
  geom_col()+
  scale_x_discrete(breaks=factor(seq(1, 50, 7)))+
  labs(x="Maximum number", y="Count")
```

Similar to the previous discussion, the smallest possible maximum number that could occur would be
7, resulting in a draw of seven consecutive numbers (1 to 7). This means that the x-axis could be
shorted to begin at 7.


### Which numbers have NEVER been in the minimum/maximums?

(Keeping in mind that we have data for roughly half of 2019, the full data for 2020, and roughly
half of 2021.)

Using the "completed" data, we can obtain the numbers which have never been the minimum/maximum by
filtering for numbers with a count of zero.

```{r}
never_min <- min_counts_complete %>%
  filter(n==0) %>%
  pull(num1)

never_min
```

It is unsurprising that 44 has never been the minimum, since if 44 ever was the minimum, the seven
winning numbers would be the consecutive numbers between 44 and 50.

```{r}
never_max <- max_counts_complete %>%
  filter(n==0) %>%
  pull(num7)

never_max
```

From these results, it looks like we should pick at least one number between 1 and 15 (and or even
between 17 and 23), and at least one number between 30 and 50!


### Most frequent minimum and maximum - by year

```{r}
min_by_year <- lottomax_wide %>%
  group_by(year, num1) %>%
  summarise(freq=n(), .groups="drop") %>%
  group_by(year) %>%
  slice_max(order_by=freq, n=3) %>%
  ungroup()

min_by_year
```

```{r}
ggplot(min_by_year, aes(x=factor(num1), y=freq))+
  geom_col()+
  facet_wrap(~year, scales="free_x")+
  labs(x="Minimum number", y="Count")
```

The x-axis of this plot is ordered by the minimum number drawn and not by the counts. But we
*can* order by counts if we want to. We will make use of the functions `reorder_within` and
`scale_x_reordered` from the `tidytext` package.

```{r}
min_by_year %>%
  mutate(num1_reordered = tidytext::reorder_within(num1, by=desc(freq), within=year)) %>%
  ggplot(aes(x=num1_reordered, y=freq))+
  geom_col()+
  tidytext::scale_x_reordered()+
  facet_wrap(~year, scales="free_x")+
  labs(x="Minimum number", y="Count")
```

The above can be repeated using the maximum number.

```{r}
max_by_year <- lottomax_wide %>%
  group_by(year, num7) %>%
  summarise(freq=n(), .groups="drop") %>%
  group_by(year) %>%
  slice_max(order_by=freq, n=3) %>%
  ungroup()

max_by_year
```

```{r}
ggplot(max_by_year, aes(x=factor(num7), y=freq))+
  geom_col()+
  facet_wrap(~year, scales="free_x")+
  labs(x="Maximum number", y="Count")
```

```{r}
max_by_year %>%
  mutate(num7_reordered = tidytext::reorder_within(num7, by=desc(freq), within=year)) %>%
  ggplot(aes(x=num7_reordered, y=freq))+
  geom_col()+
  tidytext::scale_x_reordered()+
  facet_wrap(~year, scales="free_x")+
  labs(x="Maximum number", y="Count")
```


### Summary statistics for minimum and maximum

```{r}
lottomax_wide %>%
  select(num1, num7) %>%
  sapply(summary)
```


### Distribution of the sum of the seven main numbers

```{r}
draw_sums <- lottomax_wide %>%
  select(date, contains("num")) %>%
  rowwise() %>%
  mutate(total = sum(c_across(contains("num"))))

draw_sums

ggplot(draw_sums, aes(x=total))+
  geom_histogram(bins=9)+
  labs(x="Sum of seven main numbers", y="Frequency")
```


### Sum of the seven main numbers with respect to time

I want to make an additional variable so that it will be easier to distinguish neighbouring months.
Two colours will be used &mdash; one for even months and one for odd months.

```{r}
draw_sums <- draw_sums %>%
  mutate(month_colour = factor(month(date) %% 2))
```

```{r}
ggplot(draw_sums, aes(x=date, y=total))+
  geom_point(aes(colour=month_colour), show.legend=FALSE) +
  geom_line(alpha=0.5)+
  scale_colour_manual(values=c("#07a707", "#4000c4"))+
  scale_x_date(date_breaks="3 months", date_labels="%b '%y")+
  labs(x="Date", y="Sum of seven main numbers")
```

This graph looks a bit squished. Might look better if we facet by year.

```{r}
draw_sums <- draw_sums %>%
  mutate(year = factor(year(date)))

ggplot(draw_sums, aes(x=date, y=total))+
  geom_point(aes(colour=month_colour), show.legend=FALSE)+
  geom_line(alpha=0.5)+
  scale_colour_manual(values=c("#07a707", "#4000c4"))+
  scale_x_date(date_breaks="3 months", date_labels="%b")+
  facet_wrap(~year, nrow=3, scales="free_x")+
  labs(x="Date", y="Sum of seven main numbers")
```


### Minimums and maximums over time

```{r}
min_max <- lottomax_wide %>%
  select(year, date, num1, num7) %>%
  mutate(month_colour = factor(month(date) %% 2))

ggplot(min_max, aes(x=date))+
  geom_point(aes(y=num1, colour=month_colour), show.legend=FALSE)+
  geom_line(aes(y=num1), alpha=0.5)+
  geom_point(aes(y=num7, colour=month_colour), show.legend=FALSE)+
  geom_line(aes(y=num7), alpha=0.5)+
  scale_colour_manual(values=c("#07a707", "#4000c4"))+
  scale_x_date(date_breaks="3 months", date_labels="%b '%y")+
  labs(x="Date", y="Number drawn")
```

Again, this looks a bit squished. Facet by year.

```{r}
ggplot(min_max, aes(x=date))+
  geom_point(aes(y=num1, colour=month_colour), show.legend=FALSE)+
  geom_line(aes(y=num1), alpha=0.5)+
  geom_point(aes(y=num7, colour=month_colour), show.legend=FALSE)+
  geom_line(aes(y=num7), alpha=0.5)+
  scale_colour_manual(values=c("#07a707", "#4000c4"))+
  scale_x_date(date_breaks="3 months", date_labels="%b")+
  facet_wrap(~year, nrow=3, scales="free_x")+
  labs(x="Date", y="Number drawn", caption="Minimum and maximum number drawn by date")
```


## Long data

```{r}
lottomax_long <- readRDS("./data/lottomax_long.RDS")

lottomax_long
```


### Most frequent minimums and maximums - overall

```{r}
min_counts <- lottomax_long %>%
  filter(role == "num1") %>%
  count(number) %>%
  arrange(desc(n))
```
